:ok_hand:♻️  Remove ts-ignore's  by extracting components types into their own TS files
