+++ src/components/HdAlertTypes.ts
@@ -1,0 +1,22 @@
+import {
+  checkCircle as successIcon,
+  info as infoIcon,
+  warning as warningIcon,
+  error as errorIcon,
+} from 'homeday-assets';
+
+// The first item is the default
+export const TYPES = {
+  info: {
+    icon: infoIcon,
+  },
+  success: {
+    icon: successIcon,
+  },
+  warning: {
+    icon: warningIcon,
+  },
+  error: {
+    icon: errorIcon,
+  },
+};
