+++ src/components/HdBadgeTypes.ts
@@ -1,0 +1,7 @@
+export const TYPES = {
+  DEFAULT: '',
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  TERTIARY: 'tertiary',
+  DISABLED: 'disabled',
+};
