+++ src/components/HdLinkTypes.ts
@@ -1,0 +1,4 @@
+export const TYPES = [
+  'primary',
+  'secondary',
+];
