+++ src/components/buttons/HdButtonTypes.ts
@@ -1,0 +1,7 @@
+export const TYPES = {
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  TERTIARY: 'tertiary',
+  FLAT: 'flat',
+  GHOST: 'ghost',
+};
