+++ src/components/notifications/HdNotificationsTypes.ts
@@ -1,0 +1,6 @@
+export const TYPES = {
+  ERROR: 'error',
+  NOTIFICATION: 'notification',
+  INFO: 'info',
+  SUCCESS: 'success',
+};
